a better version of the utilities used for the tests
